[Ocp] - Copy kubeconfig file to cluster root dir
When deploying ocp cluster by using "ocp" role, in order to keep up with
the consistency with other roles and ease clusters import flow with
"acm_import_cluster" role, copy the kubeconfig file of each deployed
cluster to the root of its directory.

This file will be fetched by the "acm_import_cluster" to perform the
import of the cluster into the Hub.