[ACM_HIVE_CLUSTER] - Comment out default var
In "acm_hive_cluster", the default variable "acm_import_clusters" is
used to define the clusters that should be imported.
But if this role is used in a general playbook that may need to skip the
execution of the role, the definition of the variable with unexisting
values may fail the execution of the entire playbook.

Comment out the default variable, so it will be defined only when
explicitly used.