[MANAGED_OPENSHIFT] - Modify ROSA cluster version deploy (#60)
Modify ROSA cluster version selection for deployment.
With the change, the user should be able to define the requested version
to be deployed instead of searching for specific version and defining
it prior the deployment.

In the new flow, the requested version defined by the user, for example
- 4.15, will be automatically searched across available versions and the
last availabe version will be selected for ROSA cluster deployment.