[MANAGED_OPENSHIFT] - Fix ROSA admin user creation
During the flow of admin user creation for ROSA cluster, admin
credentials are fetched.
The flow always uses the latest ROSA binary to ensure the is no api
break.

In the latest ROSA binary, during admin user creation, the following
warning message appears:
"Region flag will be removed from this command in future versions"

This warning breaks the credentials fetch flow after the admin creation.
Change the module and the command to avoid such an issue.